requirements.txt
anthropic
langgraph
langchain
langchain-anthropic
streamlit
plotly
pypdf
sentence-transformers
chromadb
python-dotenv